Final Warning pass (#99)
* started on basic type errors

* Done. For now

* remove libgfxd binary

* Fix Clang Warnings

* Fix2

* Fix rollback of #92 and missed warning in ZCutscene.cpp

* ?

* add stdexcept to ZResource

* Fix make clean

* PR fixes

* Fix make file stuff

* Remove many more errors and remove -Wall

* Turns out Globals was used

* Final pass

+++ ZAPD/Main.cpp
@@ -41,7 +41,7 @@ void ErrorHandler(int sig)
 	size = backtrace(array, 4096);
 	symbols = backtrace_symbols(array, 4096);
 
-	for (int i = 1; i < size; i++)
+	for (size_t i = 1; i < size; i++)
 	{
 		//size_t len = strlen(symbols[i]);
 		cout << symbols[i] << "\n";
